feat: reservation additional

+++ src/main/java/com/reservation/reservation/service/ReservationServiceImpl.java
@@ -2,6 +2,7 @@ package com.reservation.reservation.service;
 
 import com.reservation.auth.entity.User;
 import com.reservation.auth.repository.UserRepository;
+import com.reservation.auth.service.AuthService;
 import com.reservation.common.exception.ResourceNotFoundException;
 import com.reservation.common.exception.RoomUnavailableException;
 import com.reservation.reservation.dto.request.CreateReservationRequest;
@@ -19,9 +20,13 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
+import com.reservation.auth.entity.UserRole;
+import org.springframework.security.access.AccessDeniedException;
 
 import java.math.BigDecimal;
 import java.time.Duration;
+import java.time.LocalDate;
+import java.time.LocalTime;
 import java.util.List;
 
 @Service
@@ -32,6 +37,7 @@ public class ReservationServiceImpl implements ReservationService {
     private final RoomRepository roomRepository;
     private final UserRepository userRepository;
     private final ReservationMapper reservationMapper;
+    private final AuthService authService;
 
     @Override
     public ReservationResponse create(CreateReservationRequest request) {
@@ -50,6 +56,19 @@ public class ReservationServiceImpl implements ReservationService {
                     "Room is not available for reservation."
             );
         }
+
+        validateReservationTime(
+                request.getStartTime(),
+                request.getEndTime()
+        );
+        validateBusinessHours(
+                request.getStartTime(),
+                request.getEndTime());
+        validateReservationConflict(
+                room,
+                request.getReservationDate(),
+                request.getStartTime(),
+                request.getEndTime());
 
         Reservation reservation = reservationMapper.toEntity(request);
 
@@ -85,33 +104,60 @@ public class ReservationServiceImpl implements ReservationService {
                                 "id",
                                 id));
 
-        reservationMapper.updateEntity(request, reservation);
+        validateReservationAccess(reservation);
 
+        // Tentukan room yang akan digunakan
+        Room room = reservation.getRoom();
 
-        // update room
         if (request.getRoomId() != null &&
-                !request.getRoomId().equals(reservation.getRoom().getId())) {
+                !request.getRoomId().equals(room.getId())) {
 
-            Room room = roomRepository.findById(request.getRoomId())
+            room = roomRepository.findById(request.getRoomId())
                     .orElseThrow(() ->
                             new ResourceNotFoundException(
                                     "Room",
                                     "id",
                                     request.getRoomId()));
-
-            reservation.setRoom(room);
         }
 
+        // Gunakan nilai baru jika dikirim, jika tidak gunakan nilai lama
+        LocalDate reservationDate =
+                request.getReservationDate() != null
+                        ? request.getReservationDate()
+                        : reservation.getReservationDate();
 
-        // Hitung ulang total harga jika jam berubah
-        long hours = Duration.between(
-                reservation.getStartTime(),
-                reservation.getEndTime()
-        ).toHours();
+        LocalTime startTime =
+                request.getStartTime() != null
+                        ? request.getStartTime()
+                        : reservation.getStartTime();
+
+        LocalTime endTime =
+                request.getEndTime() != null
+                        ? request.getEndTime()
+                        : reservation.getEndTime();
+
+        // Validasi
+        validateReservationTime(startTime, endTime);
+
+        validateBusinessHours(startTime, endTime);
+
+        validateReservationConflict(
+                reservation.getId(),
+                room,
+                reservationDate,
+                startTime,
+                endTime
+        );
+
+        // Baru update entity
+        reservationMapper.updateEntity(request, reservation);
+        reservation.setRoom(room);
+
+        // Hitung ulang total harga
+        long hours = Duration.between(startTime, endTime).toHours();
 
         reservation.setTotalPrice(
-                reservation.getRoom()
-                        .getPrice()
+                room.getPrice()
                         .multiply(BigDecimal.valueOf(hours))
         );
 
@@ -129,6 +175,10 @@ public class ReservationServiceImpl implements ReservationService {
                                 "Reservation",
                                 "id",
                                 id));
+
+
+        validateReservationAccess(reservation);
+
 
         return reservationMapper.toResponse(reservation);
     }
@@ -158,6 +208,7 @@ public class ReservationServiceImpl implements ReservationService {
             Long id,
             ReservationStatusRequest request
     ) {
+        validateAdmin();
 
         Reservation reservation = reservationRepository.findById(id)
                 .orElseThrow(() ->
@@ -173,18 +224,20 @@ public class ReservationServiceImpl implements ReservationService {
         return reservationMapper.toResponse(reservation);
     }
 
-    @Override
-    public void delete(Long id) {
+        @Override
+        public void delete(Long id) {
 
-        Reservation reservation = reservationRepository.findById(id)
-                .orElseThrow(() ->
-                        new ResourceNotFoundException(
-                                "Reservation",
-                                "id",
-                                id));
+            Reservation reservation = reservationRepository.findById(id)
+                    .orElseThrow(() ->
+                            new ResourceNotFoundException(
+                                    "Reservation",
+                                    "id",
+                                    id));
 
-        reservationRepository.delete(reservation);
-    }
+            validateReservationAccess(reservation);
+
+            reservationRepository.delete(reservation);
+        }
 
     /**
      * Ambil user yang sedang login
@@ -202,5 +255,113 @@ public class ReservationServiceImpl implements ReservationService {
                                 "User",
                                 "email",
                                 email));
+    }
+
+    private void validateReservationAccess(Reservation reservation) {
+
+        User currentUser = getCurrentUser();
+
+        // Admin boleh mengakses semua reservation
+        if (currentUser.getRole() == UserRole.ADMIN) {
+            return;
+        }
+
+        // Customer hanya boleh mengakses reservation miliknya
+        if (!reservation.getCustomer().getId().equals(currentUser.getId())) {
+            throw new AccessDeniedException(
+                    "You are not allowed to access this reservation"
+            );
+        }
+    }
+
+    private void validateAdmin() {
+
+        User currentUser = getCurrentUser();
+
+        if (currentUser.getRole() != UserRole.ADMIN) {
+            throw new AccessDeniedException(
+                    "Only admin can update reservation status"
+            );
+        }
+    }
+
+    private void validateReservationTime(
+            LocalTime startTime,
+            LocalTime endTime
+    ) {
+
+        if (!endTime.isAfter(startTime)) {
+            throw new IllegalArgumentException(
+                    "End time must be after start time."
+            );
+        }
+    }
+
+    private static final LocalTime OPEN_TIME =
+            LocalTime.of(14, 0);
+
+    private static final LocalTime CLOSE_TIME =
+            LocalTime.of(22, 0);
+
+    private void validateBusinessHours(
+            LocalTime startTime,
+            LocalTime endTime
+    ) {
+
+        if (startTime.isBefore(OPEN_TIME)
+                || endTime.isAfter(CLOSE_TIME)) {
+
+            throw new IllegalArgumentException(
+                    "Reservation time is outside business hours."
+            );
+        }
+    }
+
+    private void validateReservationConflict(
+            Room room,
+            LocalDate reservationDate,
+            LocalTime startTime,
+            LocalTime endTime
+    ) {
+
+        boolean conflict =
+                reservationRepository
+                        .existsByRoomAndReservationDateAndStartTimeLessThanAndEndTimeGreaterThan(
+                                room,
+                                reservationDate,
+                                endTime,
+                                startTime
+                        );
+
+        if (conflict) {
+            throw new IllegalArgumentException(
+                    "Selected time slot is already reserved."
+            );
+        }
+    }
+
+    private void validateReservationConflict(
+            Long reservationId,
+            Room room,
+            LocalDate reservationDate,
+            LocalTime startTime,
+            LocalTime endTime
+    ) {
+
+        boolean conflict =
+                reservationRepository
+                        .existsByRoomAndReservationDateAndStartTimeLessThanAndEndTimeGreaterThanAndIdNot(
+                                room,
+                                reservationDate,
+                                endTime,
+                                startTime,
+                                reservationId
+                        );
+
+        if (conflict) {
+            throw new IllegalArgumentException(
+                    "Selected time slot is already reserved."
+            );
+        }
     }
 }
